% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMeM_reml.R
\name{MMeM_reml}
\alias{MMeM_reml}
\title{Multivariate REML Method}
\usage{
MMeM_reml(fml, data, factor_X, T.start, E.start, maxit = 50,
  tol = 1e-09)
}
\arguments{
\item{fml}{a two-sided linear formula object describing both the fixed-effects and random-effects  parts  of  the  model,
 with  the  response  on  the  left  of  a ~ operator. For univariate response, put variable name directly; for multivariate responses
 combine variables using concatenate operator, for example, for bivariate responses, c(var1, var2).  The predictor terms are separated  by + operators,  on  the  right.   Random-effects  terms  are
distinguished by vertical bars '|' separating expressions for design matrices from grouping factors.}

\item{data}{data frame containing the variables named in formula.}

\item{factor_X}{(logical) indicating whether predictor is a factor or continuous. By default is TRUE}

\item{T.start}{the starting matrix for the variance covariance matrix of the block random effects, it has to be positive definite q by q symmetric matrix.}

\item{E.start}{the starting matrix for the variance covariance matrix of the block random effects, it has to be positive definite q by q symmetric matrix.}

\item{maxit}{the maximum number of iterations}

\item{tol}{the convergence tolerance}
}
\value{
the output \code{T.estimates} is he estimated matrix of the variance covariance matrix of the block random effects;
\code{E.estimates} is the estimated matrix of the variance covariance matrix of the residuals; \code{VCOV} is the asymptotic
dispersion matrix of the estimated variance covariance components.
}
\description{
Estimating the variance components under the multivariate mixed effects model using REML methods
}
\details{
The model supports multivariate mixed effects model for one-way randomized block design with equal design matrices:
\deqn{Y = \left(I_q x X\right)B + \left(I_q x Z \right)U + E}

The model also supports simple OLS multivariate regression:
\deqn{y = Xb + Zu+ e}
}
\examples{
\dontrun{# simulating data
z1 = c(rep(1,45))
z2 = c(rep(1,37))
z3 = c(rep(1,40))
x1 = rbind(matrix(rep(c(1,1,0),14), 14,3, byrow = TRUE),matrix(rep(c(1,0,1),18), 18,3,
byrow = TRUE),matrix(rep(c(1,0,0),13), 13,3, byrow = TRUE))
x2 = rbind(matrix(rep(c(1,1,0),11), 11,3, byrow = TRUE),matrix(rep(c(1,0,1),13), 13,3,
byrow = TRUE),matrix(rep(c(1,0,0),13), 13,3, byrow = TRUE))
x3 = rbind(matrix(rep(c(1,1,0),11), 11,3, byrow = TRUE),matrix(rep(c(1,0,1),15), 15,3,
byrow = TRUE),matrix(rep(c(1,0,0),14), 14,3, byrow = TRUE))
Z = as.matrix(bdiag(z1, z2, z3))
X = rbind(x1,x2,x3)
B<-matrix(c(2, 3, 1, 2.5, 3.5, 1.5),3,2)
T.true<-matrix(c(100,5,5,10),2,2)
E.true<-matrix(c(60,13,13,40),2,2)
s = ncol(Z)
h = ncol(X)
each = c(45,37,40)
Z_vec = c()
for(i in 1:3){
  Z_vec = c(Z_vec, rep(i, each[i]))
}
X_vec = X[,2]+X[,3]*2
U<-mvrnorm(n = s, mu=c(0,0), Sigma=T.true)
e<-mvrnorm(n = N, mu=c(0,0), Sigma=E.true)
Y<-X\%*\%B+Z\%*\%U+e

T.start = matrix(c(10,5,5,15),2,2)
E.start = matrix(c(6,1,1,3),2,2)
data = as.data.frame(cbind(Y, X_vec, Z_vec))
results = MMeM_reml(c(V1,V2) ~ X_vec + (1|Z_vec), data,
factor_X = TRUE, T.start, E.start, maxit = 10)
}

}
\references{
Meyer, K. A. R. I. N. "Maximum likelihood estimation of variance components for a multivariate mixed model with equal design matrices." Biometrics 1985: 153-165.
}
