% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{MMeM_terms}
\alias{MMeM_terms}
\title{parses formulas to creates model matrices}
\usage{
MMeM_terms(fml, data, factor_X)
}
\arguments{
\item{fml}{a two-sided linear formula object describing both the fixed-effects and random-effects  parts  of  the  model,
 with  the  response  on  the  left  of  a ~ operator. For univariate response, put variable name directly; for multivariate responses
 combine variables using concatenate operator, for example, for bivariate responses, c(var1, var2).  The predictor terms are separated  by + operators,  on  the  right.   Random-effects  terms  are
distinguished by vertical bars '|' separating expressions for design matrices from grouping factors.}

\item{data}{data frame containing the variables named in formula.}

\item{factor_X}{(logical) indicating whether predictor is a factor or continuous. By default is TRUE}
}
\description{
parses formulas to creates model matrices
}
