% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\docType{data}
\name{mydata}
\alias{mydata}
\title{simulated bivariate data}
\usage{
data(mydata)
}
\description{
This is a simulated data with 2 dependent variables and one fixed effects
and one random effects
}
\details{
Data document
}
\examples{
\dontrun{data(mydata)}
}
\keyword{datasets}
