% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMeM.R
\docType{data}
\name{simdata}
\alias{simdata}
\title{simulated bivariate data}
\usage{
data(simdata)
}
\description{
This is a simulated data with 2 dependent variables and one fixed effects
and one random effects
}
\details{
simulated datasets
}
\keyword{datasets}
