% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMeM.R
\docType{package}
\name{MMeM}
\alias{MMeM}
\alias{MMeM-package}
\title{MMeM: Estimating the variance covariance components of the multivariate mixed effects model}
\description{
Analyzing data under multivariate mixed effects model using multivariate REML and multivariate Henderson3 methods.
}
\author{
Luyao Peng \email{luyaopeng.cn@gmail.com}

Rui Yang \email{ray.cn.us@gmail.com}
}
