% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMeM.R
\docType{data}
\name{mydata}
\alias{mydata}
\title{simulated bivariate data}
\usage{
data(mydata)
}
\description{
This is a simulated data with 2 dependent variables and one fixed effects
and one random effects
}
\keyword{datasets}
